tdf#148384: Really leave mail merge results dialog open upon errors
Restructuring mail merge in 5.2 cycle had broken the dialog keeping
open when errors happened during sending. Later, in 2017, in commit
1aae7a14e68973ee81b6d2767ab6382055eb604a
  Author Alex McMurchy1917 <mcmurchy1917-libreoffice@yahoo.co.uk>
  Date   Thu Aug 24 14:34:13 2017 +0100
    tdf#103919 mailmerge: prevent premature end of emailing

SwSendMailDialog::AllMailsSent was changed in attempt to restore that
expected behavior, but the condition was incorrect:
m_nSendCount == m_nExpectedCount means "all mails processed", not
"everything went without errors".

So fix this by considering the error count when deciding to close the
dialog.

Change-Id: I98d61165803deaf0da985167ba0152e6f50f9246
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/132563
Tested-by: Mike Kaganski <mike.kaganski@collabora.com>
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>