tdf#148380 writerfilter: Import AUTHOR as FIXEDFLD
This patch depends on patches for tdf#147861
and will require a follow-up export patch that handles exporting
this FIXEDFLD (instead of just dumping the name of the field).

This was being imported as plain text because of exception
writerfilter/source/dmapper/DomainMapper_Impl.cxx:6950:
Exception in CloseFieldCommand() com.sun.star.beans.UnknownPropertyException
message: "Unknown property: FullName at
/persistent/git/libreoffice2/sw/source/core/unocore/unofield.cxx:2156"

Additionally, similar to tdf#134592, the create author will always
be constant, so mark that as fixed. (This helps because in MS Word,
the user can modify the field contents to be anything he wants,
and it is only updated when F9 is pressed on the selected field.
Thus, we might NOT want to show the real DocInfo.CreateAuthor
as the unit test demonstrates.)

Change-Id: Ie014c273d491fa5bbd9b9ecedc69702d0ae8d60c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/132661
Tested-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>