tdf#148380 docx export: use content of unknown fields, not name
This used to be GetCntnt, but then out of the blue in a huge
squashed commit, it changed to GetFieldName. So I assume that
was just debugging code that got accidentally committed.
It just doesn't make any sense to export the field name
as the plain text content instead of the field's value.

commit 25bcbfa0e929ee8539443df0c1b58313875fab27
Author: Michael Meeks on Mon Mar 14 16:51:14 2011 +0000
    Merge commit 'ooo/DEV300_m101' into integration/dev300_m101

NOTE: DOC also does a .replace(0x0A, 0x0B),
//replace LF 0x0A with VT 0x0B
//#i19604# convert hard line breaks inside fields to word equivalent

After taking care of SAVEDATE, PRINTDATE, I tested this with
assert(rInfos.pField->GetFieldName() == "Sender");
and found no other examples.

Change-Id: Ie24b8c176fe6df4015be221c87aa249fe9b8ce92
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/132666
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>