tdf#132475 writerfilter: use proper date-field defaults
When a field doesn't specify the format for a date,
the system is free to implement it as it chooses.
However, the user tends to disagree with that,
so lets try to create a MS-similar default for
these dates - since LO defaults to only displaying
a date without any time component.

For example: PRINTDATE \* MERGEFORMAT
doesn't tell us how to display the date.

MS Word uses different defaults depending on
the language of the document. And that only make sense.
For example, I noticed that of course en-GB's format
is dd/MM/yyyy instead of en-US's M/d/yyyy.

As a documentation example: 17.16.5.47 PRINTDATE
Syntax: PRINTDATE [ switches ]
Description: Retrieves the date and time on which the
document was last printed, as recorded in the
LastPrinted element of the Core File Properties part.

Switches: Zero or one date-and-time-formatting-switch
and zero or one of the following field-specific-switches.

PRINTDATE (without any formatting switches)
   the results are:
1/6/2006 2:58:00 PM

Change-Id: I41f3bdc155bd8cdc74177e4626b31ece9b47e16b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/133208
Tested-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>