tdf#148380 sw fieldui: always allow turning off "Fixed Content"
As we add more and more MS fields as fixed,
some of them are set so the UI to manage that flag
is disabled. While it doesn't necessarily make
much sense to allow a user to SET these fields
as fixed (so don't ALWAYS enable that flag),
at least allow it to be turned off if it somehow
did get enabled.

Change-Id: I331c871fa7a3e40a7a0154ab4d7d68c67d9c1dad
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/133086
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>