sfx2: fix utterly confusing bools in SfxViewFrame::ExecReload_Impl()
bPhysObjIsYounger variable is actually true if aPhysObj is *older* than
aMedObj.

Due to the "bNeedsReload = true" that was added in commit
01ac5a64a6a5bf107aed185ec957cd7f67293991 i had first thought that commit
73919cccceecd301f9f7a4c404f02e379d8f05cc is at fault and inverted the
condition - but not so, its only crime is mis-naming the variable and
the "bNeedsReload = true" is actually dead code because this branch
always returns before any use of bNeedsReload further below.

Meanwhile, commit f1035312f80746fdf4a7a04039db73b27d1387c3 actually did
invert the condition, but that was in 2004 and presumably intentional.

Change-Id: Ie342163a1de434f99c521e97711157a76191da0b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/133029
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>