fix an off-by-one error in GetEmptyLinesInBlock()
The function has been there since the initial commit and is not
documented, but I think it counts the shortest amount of empty
cells in the given area starting from the direction given. And
AFAICT the off-by-one error was there since the initial commit,
when it returned one less if the entire area was empty and
the direction was vertical (horizontal was fine). And
ScHTMLExport::FillGraphList() even was adjusted for this until
my recent commit changing that code). But then 6cba9f61c3ddbde4df
added a shortcut for unallocated columns that didn't have
the error. And the error even got corrected during the rewrite
in 6fa5f6671f65d591fc8f3106ec0a4f0ea97b7d51, but then
a0f1dd2c402275ba7e829d934953b51027b702ba reverted that. Anyway,
so fix this, I think all the relevant code should(?) now work
properly.

Change-Id: I194691f7276a1cea75945de05cb0dda2cdca859a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/133319
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>