do not duplicate code for ScCountIfCellIterator
8644c9c6f19f98fda1e58224619b074715fc0a5e improved COUNTIF performance
by copy&pasting the generic query iterator and then basically removing
what was not necessary. Which is in general a good way to improve
the performance, except for the copy&paste part, as the code has
already started to diverge (e.g. 1b89ef8ef085faae1a34c2074777d6df1f5).

So instead make the shared code into a template and reuse that from
specific code. This has exactly the same performance as the copy&paste
way.

Change-Id: I168319a1f4273bafc8c0742a114dafbf433968bb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/133324
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>