Drop a bit more special-casing for Chart's sidebar property deck
Introduced in commit 5b89f686093152546a3bdcf5307da43c6bd92381
  Author Muhammet Kara <muhammet.kara@collabora.com>
  Date   Mon Nov 25 21:55:31 2019 +0300
    tdf#94288: Show chart props sidebar on activation

The problem was that ChartController::attachFrame, that called
SelectionChangeHandler::selectionChanged notification, did that
*prior* to setting its m_xFrame - and the notification failed
in ContextChangeEventMultiplexer::NotifyContextChange, that
checks the frame first. That prevented the proper context (with
correct application and context names) to arrive to listeners,
and the sidebar didn't update properly.

Changing the order of the calls should fix the original problem.

Change-Id: I9da8465af2ee4ed1f8eabed1c65d1c318f81a3f9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/133326
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>