tdf#141391 sw: don't paste as nested table in first cell paragraph
Paste table content overwrote cells of the row(s) only if the text
cursor was at the beginning of the table cell, otherwise the table
cells on the clipboard were inserted as a nested table.

This was a UX regression from commit 78bb8dee17197ff38a6b3a8bffcfb4e54eff026a
"fdo#37156 insert table copy as nested table in non-starting cell position"
especially when the user clicked not exactly at the beginning of a
cell, which containing a 1-line text or data.

Since commit 46ba69c278caf6111abe1d0b8eec8d354c848677
"tdf#37156 Writer menu: Paste as Nested table" it's possible to
force nesting (but not overwriting yet), this commit revert partially
commit 78bb8dee17197ff38a6b3a8bffcfb4e54eff026a: if the text cursor
is there in the first paragraph of the cell, Paste table content
overwrites the row, not embedding a nested table in the cell
at the cursor position. This change results also better interoperability
with the existing document editors.

Note: table and text selection were checked with the change, too.

Details:

Heuristics to allow copying table rows or nesting tables
without using Edit -> Paste Special -> Paste as Nested Table:

At "table selection" (i.e. when cell(s) completely selected),
or if the text selection starts in the first paragraph, or if
there is no selection and the text cursor is there in the first
paragraph, overwrite content of the cell(s). Otherwise insert
a nested table, i.e. if nothing selected and the cursor is not
in the first paragraph, or the selected text doesn't contain
the first paragraph of the cell.

Change-Id: I7746c6a464123bef6fb7dbff415ff0414e48365d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/133377
Tested-by: Jenkins
Reviewed-by: László Németh <nemeth@numbertext.org>