tdf#141835 handle SHOWWINDOWMODE_END when resuming the slide show
The problem which is fixed with this commit is, that the slide show ends when the presention screen has shown "Click to exit" after the last slide, and then the slide show was moved to a specific slide. e.g. via Remote App or Presenter Console. In this case the slide show should be resumed and the slide should be displayed. Instead the slide show was ended.

Instead of only handling the BLANK mode for resuming a slide show also the END mode is now inlcuded to resume the slide show.

Change-Id: I78e00dc882b2d7af8b7a6b215aaac122bbd9ead9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/133337
Tested-by: Jenkins
Reviewed-by: Thorsten Behrens <thorsten.behrens@allotropia.de>